refactor(automation): split renewal_reminder's catalog schema into its own file
automations_catalog.go crossed the 500-line product ceiling once #706's
schema/validator additions landed. Split into
automations_catalog_renewal.go, mirroring the identical size-not-concept
split automations_preview_renewal.go already made for the preview side
of the same catalog key. Also collapsed the "integer"/"minimum"/
"default" schema-property literal and the "must be an integer" refusal
into shared helpers (intParamProperty, reasonMustBeInteger) reused by
singleIntParamSchema, renewalReminderSchema, and leadRoutingSchema — the
duplication the strict lint gate's goconst check caught once a fourth
call site crossed its threshold.

No behavior change; make check-backend green.
